+++ src/DotPulsar/Internal/Abstractions/Process.cs
@@ -35,6 +35,7 @@ public abstract class Process : IProcess
     }
 
     public Guid CorrelationId { get; }
+    protected ulong? TopicEpoch { get; private set; }
 
     public void Start()
         => CalculateState();
@@ -56,6 +57,10 @@ public abstract class Process : IProcess
                 ChannelState = ChannelState.ClosedByServer;
                 break;
             case ChannelConnected _:
+                ChannelState = ChannelState.Connected;
+                break;
+            case ProducerChannelConnected producerChannelConnected:
+                TopicEpoch = producerChannelConnected.TopicEpoch;
                 ChannelState = ChannelState.Connected;
                 break;
             case ChannelDeactivated _: