+++ src/DotPulsar/Internal/NotReadyChannel.cs
@@ -34,8 +34,6 @@ public sealed class NotReadyChannel<TMessage> : IConsumerChannel<TMessage>, IPro
     public ValueTask<IMessage<TMessage>> Receive(CancellationToken cancellationToken = default)
         => throw GetException();
 
-    public ulong? TopicEpoch { get => null; }
-
     public Task Send(MessageMetadata metadata, ReadOnlySequence<byte> payload, TaskCompletionSource<BaseCommand> responseTcs, CancellationToken cancellationToken)
         => throw GetException();