+++ src/DotPulsar/Internal/Producer.cs
@@ -74,7 +74,7 @@ public sealed class Producer<TMessage> : IProducer<TMessage>, IRegisterEvent
         };
         _attachTraceInfoToMessages = options.AttachTraceInfoToMessages;
         _sequenceId = new SequenceId(options.InitialSequenceId);
-        _state = new StateManager<ProducerState>(ProducerState.Disconnected, ProducerState.Closed, ProducerState.Faulted);
+        _state = CreateStateManager();
         ServiceUrl = serviceUrl;
         Topic = options.Topic;
         _isDisposed = 0;
@@ -146,27 +146,32 @@ public sealed class Producer<TMessage> : IProducer<TMessage>, IRegisterEvent
                 switch (state)
                 {
                     case ProducerState.Connected:
-                        ++connectedProducers;
-                        waitingForExclusive[i] = false;
+                        if (waitingForExclusive[i])
+                            waitingForExclusive[i] = false;
+                        else
+                            ++connectedProducers;
                         break;
                     case ProducerState.Disconnected:
                         --connectedProducers;
+                        waitingForExclusive[i] = false;
                         break;
                     case ProducerState.WaitingForExclusive:
+                        ++connectedProducers;
                         waitingForExclusive[i] = true;
                         break;
+                    case ProducerState.Fenced:
                     case ProducerState.Faulted:
-                        _state.SetState(ProducerState.Faulted);
+                        _state.SetState(state);
                         return;
                 }
 
                 monitoringTasks[i] = _producers[i].OnStateChangeFrom(state, _cts.Token).AsTask();
             }
 
-            if (connectedProducers == monitoringTasks.Length)
-                _state.SetState(ProducerState.Connected);
-            else if (connectedProducers == 0 && waitingForExclusive.All(x => x != true))
+            if (connectedProducers == 0)
                 _state.SetState(ProducerState.Disconnected);
+            else if (connectedProducers == monitoringTasks.Length && waitingForExclusive.All(x => x != true))
+                _state.SetState(ProducerState.Connected);
             else if (waitingForExclusive.Any(x => x))
                 _state.SetState(ProducerState.WaitingForExclusive);
             else
@@ -181,7 +186,7 @@ public sealed class Producer<TMessage> : IProducer<TMessage>, IRegisterEvent
         var schema = _options.Schema;
         var producerAccessMode = (DotPulsar.Internal.PulsarApi.ProducerAccessMode) _options.ProducerAccessMode;
         var factory = new ProducerChannelFactory(correlationId, _processManager, _connectionPool, topic, producerName, producerAccessMode, schema.SchemaInfo, _compressorFactory);
-        var stateManager = new StateManager<ProducerState>(ProducerState.Disconnected, ProducerState.Closed, ProducerState.Faulted);
+        var stateManager = CreateStateManager();
         var initialChannel = new NotReadyChannel<TMessage>();
         var executor = new Executor(correlationId, _processManager, _exceptionHandler);
         var producer = new SubProducer(correlationId, _processManager, initialChannel, executor, stateManager, factory, _options.MaxPendingMessages);
@@ -265,7 +270,7 @@ public sealed class Producer<TMessage> : IProducer<TMessage>, IRegisterEvent
 
         try
         {
-            await InternalSend(metadata, message, true, OnMessageSent, cancellationToken).ConfigureAwait(false);
+            await InternalSend(metadata, message, true, OnMessageSent, x => tcs.TrySetException(x), cancellationToken).ConfigureAwait(false);
             return await tcs.Task.ConfigureAwait(false);
         }
         finally
@@ -275,9 +280,9 @@ public sealed class Producer<TMessage> : IProducer<TMessage>, IRegisterEvent
     }
 
     public async ValueTask Enqueue(MessageMetadata metadata, TMessage message, Func<MessageId, ValueTask>? onMessageSent = default, CancellationToken cancellationToken = default)
-        => await InternalSend(metadata, message, false, onMessageSent, cancellationToken).ConfigureAwait(false);
+        => await InternalSend(metadata, message, false, onMessageSent, cancellationToken: cancellationToken).ConfigureAwait(false);
 
-    private async ValueTask InternalSend(MessageMetadata metadata, TMessage message, bool sendOpCancelable, Func<MessageId, ValueTask>? onMessageSent = default, CancellationToken cancellationToken = default)
+    private async ValueTask InternalSend(MessageMetadata metadata, TMessage message, bool sendOpCancelable, Func<MessageId, ValueTask>? onMessageSent = default, Action<Exception>? onFailed = default, CancellationToken cancellationToken = default)
     {
         ThrowIfDisposed();
 
@@ -311,10 +316,14 @@ public sealed class Producer<TMessage> : IProducer<TMessage>, IRegisterEvent
 
                 if (task.IsFaulted || task.IsCanceled)
                 {
-                    FailActivity(task.IsCanceled ? new OperationCanceledException() : task.Exception!, activity);
+                    Exception exception = task.IsCanceled ? new OperationCanceledException() : task.Exception!;
+                    FailActivity(exception, activity);
 
                     if (autoAssignSequenceId)
                         metadata.SequenceId = 0;
+
+                    onFailed?.Invoke(exception);
+                    return;
                 }
 
                 CompleteActivity(task.Result, data.Length, activity);
@@ -375,6 +384,9 @@ public sealed class Producer<TMessage> : IProducer<TMessage>, IRegisterEvent
         if (_isDisposed != 0)
             throw new ProducerDisposedException(GetType().FullName!);
     }
+
+    private StateManager<ProducerState> CreateStateManager()
+        => new (ProducerState.Disconnected, ProducerState.Closed, ProducerState.Faulted, ProducerState.Fenced);
 
     public void Register(IEvent @event) { }
 }