+++ src/DotPulsar/Internal/ProducerChannel.cs
@@ -31,15 +31,12 @@ public sealed class ProducerChannel : IProducerChannel
     private readonly ICompressorFactory? _compressorFactory;
     private readonly byte[]? _schemaVersion;
 
-    public ulong? TopicEpoch { get; }
-
     public ProducerChannel(
         ulong id,
         string name,
         IConnection connection,
         ICompressorFactory? compressorFactory,
-        byte[]? schemaVersion,
-        ulong topicEpoch)
+        byte[]? schemaVersion)
     {
         var sendPackagePolicy = new DefaultPooledObjectPolicy<SendPackage>();
         _sendPackagePool = new DefaultObjectPool<SendPackage>(sendPackagePolicy);
@@ -48,7 +45,6 @@ public sealed class ProducerChannel : IProducerChannel
         _connection = connection;
         _compressorFactory = compressorFactory;
         _schemaVersion = schemaVersion;
-        TopicEpoch = topicEpoch;
     }
 
     public async ValueTask ClosedByClient(CancellationToken cancellationToken)